Configure travis.yml for this project
This commit is the result of several unsuccessful attempts on
configuring travis.yml file. It shouldn't be that hard, but it
always is, since you never know what is going to happen until
it actually happens.

Hope the Git gods forgive me for using a squash and then a push
force on a public repository.