Upgrade the app to Rails 5.0
This commit is the squashed result of a set of upgrades made to
this project in order to upgrade it from Rails 4.2.6 to Rails 5.0.
The following procedure was followed:

* Upgraded to Rails 4.2.7, latest patch for 4.2 series.
* Upgraded to Rails 5.0, target version I want to use.

Before and after every upgrade tests were run in order to assert
that everything was working smooth. The following issues have been
found and fixed:

* The admin panel gem is not compatible with Rails 5.0. I have
  removed it since I'm writing my own CRUDs anyway.
* Rails 5 seems to have soft-killed controller tests (see issue #1)
  and a few things that worked on Rails 4.2 are now not working
  that well.
  * Requests are now done differently.
  * Some methods such as `assigns` are now in a separate gem.