Add a playlist video organizer
The playlist video organizer lets the administrator to move up and down
videos for a particular playlist. It is not possible to delete videos
from a playlist since that assumes that it is possible to have videos
not in any playlist, and that is not possible. So, to delete a video
from a playlist it has to be added into another playlist. Do that via
the Edit Video page.