Add a topic playlist organizer
On a similar fashion, it should be possible to sort the order in which
playlists that are covered in topics appear. In this case, it is
possible to move a playlist up, move it down, but also unlink, since it
is totally possible to have playlists not in any topic.