Add a dashboard (closes #9)
The purpose of the dashboard is to let administrators and editors to            
create, update and destroy videos, topics, playlists and other data             
structures used in the website.                                                 
                                                                                
The dashboard does not have to be perfect for this initial release but          
it has to fit a few functional and security requirements, such as: it           
has to work, it has to do its job well, it has to be simple, and it             
should not let other people use the admin panel.