Update Ruby build matrix for Travis-CI
Ruby 2.3.3 has been added because this is the Ruby version the
provisioned servers run at this moment. It has support anyway so it will
stay installed in the provisioned servers for a while.

Ruby 2.4.0 has been added to test that the web application works with
recent versions of Ruby. I'm not upgrading the Ruby version on the
provisioned servers right now, but I want to know that the application
works anyway.

Ruby 2.2.3 has been removed from the build matrix because neither the
provisioning servers neither my development setup uses this version
anyway.