Convert makigas logo to a SVG image
Applying this commit, the current makigas logo rendered at the header
bar for every page in the site will be replaced from the current plain
text string rendered using a custom font, with an SVG image that
displays the same glyphs as a path.

Replacing the text logo with an SVG path removes font files and reduces
page size by inlining in less than 1 KB the actual logo for the website.