Add methods for testing video visibility
In order to decide whether a video is already visible or scheduled,
applying this commit will add a set of methods to the Video model in
order to assert the condition for a video.

* .visible? will return true if the publishing date for a video has
already been reached allowing other people to browse for the video in
the catalog.

* .scheduled? tests the opposite condition.

Additionally, a scope named .visible has been added to the Video model
allowing to query videos that are visible and discarting those who
aren't.