Introduce dotenv based configuration
After applying this commit, the current system for storing passwords and
secrets will change to dotenv and .env files. Using this approach, there
is only one single file containing the local settings for the
application.

This approach is compatible with Twelve Factor apps and may help later
to port the system to other platforms such as Docker or other kind of
application servers, since settings can come via ENV.