Replace application server with Puma
Applying this commit the application server for Rails will be replaced
with Puma. This includes some changes that have already been done
allowing Puma to technically be able to run in production using sockets
and a server such as nginx.

Note that this commit does not solve how to start Puma automatically,
something that has to be decided during provisioning using some kind of
monitoring system allowing the Puma server to be started, stopped,
restarted and possibly autostarted on crash.