Allows scheduled videos to be browsed anonymously
Apply this commit to add a field to the videos model allowing an admin
to mark a video as private. Private videos only have effect when they
are scheduled.

The old behaviour was, not to let anonymous users to view scheduled
videos. However, now they can be viewed unless this checkbox is marked.
Note that there is no checkbox at the moment. It has to be added.