Embraces Yarn, uses it for Bootstrap and jQuery
Apply this commit to add support for Yarn to this repository. Yarn was
officially supported starting from Rails 5.1. It allows you to install
JavaScript packages and it is an alternative to NPM.

This commit removes references to Bootstrap and jQuery, currently being
loaded through local services, and instead it pulls them through a
package manager and servers them locally.

They are fully integrated within the assets pipeline by default so I can
just ask Sprockets to import a CSS or a JS file located in the
node_modules/ directory.