Embraces Yarn and possibly Webpack (#50)
This commit introduces support for Yarn into the web application. Yarn
has been formally supported since Rails 5.1. I've been skeptical on
JavaScript for a long time but I think that if the changes are designed
to make this project better in the future, they should be welcomed.
    
Specifically, this commit adds a package.json for adding JavaScript
dependencies and adding them to the asset pipeline so that they can be
used when compiling stylesheets and script files.
    
This was something requested in issue #17 because some external
dependencies were being served from outside sites. The following
dependencies are now local to the project and mushed into the assets
pipeline:
    
* Bootstrap 3.
* jQuery 1.x.
* The Cookie Consent plugin.
    
Future developments may include:
    
* Include only the specific parts of Bootstrap that I'm interested in
  (for instance, don't include code for classes or features that I don't
  use such as dialogs or wells).
* Remove jQuery at all. JavaScript is powerful enough these days to do
  all that work. Some Bootstrap things depend on jQuery, such as the
  navbar, so this has to be thought.
* Maybe embrace more front-end technologies into the stack without
  rewriting the application. Rails is cool and I want to keep using
  Rails.
    
Fixes #17, because most external dependencies are now served locally.