Optimize Puma configuration file (#57)
An alternative configuration file for production environments has been
provided and located in config/puma/production.rb so that Puma
automatically uses it when booting the application in a production
server.

This configuration file increases the number of workers to make the
application server concurrent. The number of workers can be tweaked
using an environment variable and in case the application ever grows up
and needs some scaling, it can be increased as well.

The production configuration file also removes the port configuration.
Apparently setting the port, even when the application is already
binded to an UNIX socket, forces Puma to serve the application through
that port. Now the application is only binded through an UNIX socket
that can be proxied through Nginx. On development, the application is
still accessed through the TCP port as usual.