Drop :dashboard namespace for route helpers
Apply this commit to remove the :dashboard namespace on routes made
using the route helper. This means that it is not required anymore to
link to a dashboard page using dashboard_video_path, it can be done
using video_path. It is not required to use [:dashboard, :videos],
:videos can just be used.

When linking to a public asset through the dashboard, remember to use
the subdomain key in the URL helper to make sure the link points to the
appropiate subdomain.