Adds a docker-compose.yml suitable for development
This commit will add a docker-compose.yml file that can be used to run
the application in a development environment. Note that the application
has not been tested to be suitable for testing environments yet. The
docker-compose is not valid for production environments because it
mounts the entire code as a volume.