Installs Webpack and Webpacker
I guess I'm a modern front-end developer now. 🤓

Apply this commit to add support for webpack through the webpacker gem.
I've included the development version for webpacker in order to add
support for the on-demand compiling, a feature that is not currently
present in the latest released gem. If it turns out to be buggy or
unsafe for production, I'll rely to using the stable version (2.0 at the
moment), even though that requires to use the webpack-dev-server to
compile assets in development.

bin/rails webpacker:install is also executed. Therefore this commit has
everything required to start using webpack.