Port scripts to vanilla JS and drop jQuery
Apply this commit to port all the application .js files to vanilla JS,
making jQuery unnecesary for them. jquery-ujs has been replaced with the
vanilla rails-ujs added in Rails 5.1 also.

This means that makigas.es sources do not depend anymore on jQuery. It
is required for Bootstrap, though, so it's not actually being removed.
It will be removed given one of the following cases:

* Bootstrap is refactored to not depend on jQuery (vanilla JS).
* A native vanilla implementation for the navbar without relying on JS
  is developed.

The only Bootstrap component I use that depends on jQuery is the navbar.

Note that there have been a few losses with this change:

* The Read more button at the front header does not animate anymore.
  At some point it would be nice to add the animation scroll effect
  again. Clicking the button should animate the scroll.

* The front navbar is not transparent anymore in order to avoid the
  need for porting the "de-transparentize" effect when the navbar is
  toggled, as the navbar events are only sent through jQuery, not
  through native JS.