Fixes regression on timer component for videos
Apparently this broke, or was already broken. Setting the hours, minutes
and seconds was not properly triggering the JavaScript events that
update the actual duration in seconds for this field.

The JavaScript event has been replaced with a new submit event that
updates the value for the duration in seconds just before submitting the
form. Alternatively, it has been thought to simply make the backend
controller accept the duration sexagesimal components.