Some feature tests require JavaScript testing
The duration component in the video form actually requires testing that
the JavaScript code the page uses to convert the duration from
sexagesimal units to seconds actually works.

Apply this commit to add support for Poltergeist, a library that makes
use of PhantomJS to run some feature tests in a real browser with
JavaScript support, unlike the default browser used by Capybara.

This is used to fix some new regressions found in dashboard/videos.js.