Reset the app_host after running dashboard tests
If the app_host and the server_port don't get reset after running the
dashboard tests that actually need those values set, the remaining tests
will be run using the same app_host.

Forgetting about this will make all the remaining public site feature
tests to fail because, given that app_host still points to the
dashboard, they will run the tests against the dashboard instead of
against the public site.

Apply this commit to fix this by clearing the values for
Capybara.app_host and Capybara.server_port after running the tests. The
values are set to nil because after running a debug puts these seem to
be the default values (unlike default_host).