Fix dashboard front-end shenanigans (#64)
Apparently this broke, or was already broken. Setting the hours, minutes
and seconds was not properly triggering the JavaScript events that
update the actual duration in seconds for this field.

The JavaScript event has been replaced with a new submit event that
updates the value for the duration in seconds just before submitting the
form. Alternatively, it has been thought to simply make the backend
controller accept the duration sexagesimal components.

The duration component in the video form actually requires testing that
the JavaScript code the page uses to convert the duration from
sexagesimal units to seconds actually works.

In addition, this commit adds support for Poltergeist, a library that makes
use of PhantomJS to run some feature tests in a real browser with
JavaScript support, unlike the default browser used by Capybara.