Update dependencies (2017-09-28)
There are some security advisories on some JavaScript transient
dependencies this project uses. JavaScript is only used during the
deployment in order to compile via Webpack the assets, but it's
unfortunate.

This commit updates the dependencies. None of the affected dependencies
changes version, mostly because the packages these transient
dependencies are used haven't still updated yet.