Fix node_modules not present after volume mount
After removing node_modules/ in my host directory, I noticed that the
application stopped working inside the container.

Apparently, because I'm mounting . into the container, it doesn't matter
the Dockerfile runs yarn install: the node_modules directory in the
container gets hidden when the whole application directory is replaced
by the root contents. That means that it's still required to run yarn
install on the host to make node_modules be available in the container.

To avoid this, the volume-in-volume Docker trick is added in this
commit, by using information described in this article here [1].
Container's node_modules directory is remounted after binding the
application directory. This makes the original /makigas/node_modules
directory available, even if the host system doesn't have this
directory.

This trick is also used as a workaround to replace tmpfs introduced in
commit de9c93f with a local and persistent mounting point that survive
stopping and restarting the container.

[1]: http://jdlm.info/articles/2016/03/06/lessons-building-node-app-docker.html#the-node_modules-volume-trick