🐳 Add Docker build files (#59)
Apply this commit to provide a Dockerfile for building an image based on
this web application. This image is oriented towards development and
testing, which means that it is still not safe to run this image on a
production environment. This is something that I'm looking forward at
some point in the future when I consider switching the server backend.

Docker scripts can be used to make easier to develop or test features on
this web application without requiring to install a full PostgreSQL
distribution or a Ruby interpreter just for tweaking a few things.

Images are based on Alpine Linux to reduce the final image size and take
care of the process of building the application, installing
dependencies, connecting to the database, managing assets and compiling
packs.

The docker-compose.yml is designed to work with Docker Compose. It may
work with Docker Swarm if a few changes are made but Docker Swarm is
still not supported right away. I'm closely following Docker Swarm. It
looks nice but there are a few holes I'm still waiting to be closed. As
soon as it looks promising enough, I can start supporting Swarm, maybe
even in production.