S3 Storage support (#77)
This PR adds support for customizable file uploading into S3 and S3-like
storage servers. Whenever a playlist, a topic or something that has a
file field is saved, the uploaded file will be transfered into a
different server for storage. A canonical URL is generated and when the
browser downloads the web document, a full URL for accessing to the
resource through the S3 public URL is provided to the user.

Currently Paperclip is storing thumbnails and cards into the file system
by using the default uploader, using
/var/www/makigas/current/public/system as the place where to save these
data. However, there are a few advantages on using a storage server like
S3 instead of the bare server. Just to name a few:

* Assets can be served through a CDN closer to the user.
* Assets can be served throught a cache server.
* Application is easier to move, since there are less files to care about.
* Assets can be served from a cookie-less domain, making requests faster.