Don't syndicate unpublished videos
An unhandled edge case where videos that are scheduled for later
publication but not made public yet would be presented in the ATOM Feed
was corrected on this commit.

Additionally, a few includes() have been added to the video queries in
order to avoid N+1 queries when presenting the playlist title and the
topic title for a particular video.