Add a Rake task for setting Minio bucket policies
Minio buckets don't use the s3:GetObject bucket policy by default. I
believe S3 works the same. IMHO this is good because it allows
authorized users to do stuff, but it avoids anonymous users accessing
private files unless the owner of the bucket wants so.

When setting up the development environment using the docker-compose.yml
file, a Minio instance is created. But, because the bucket policy is not
set, it's not possible to view loaded files in the web browser.

You can't just use the s3_permissions key in Paperclip as described here
[1], because Minio doesn't use the ACL based permissions provided by S3.
Instead, Minio simplifies the permission settings to bucket policies.
Therefore, images won't be visible even in the following case:

    has_attached_file :s3_permissions => "public-read"

This commit adds a Rake task named makigas:minio:init_bucket that uses
the AWS-CLI to connect to the Minio instance and set the bucket policy
for the bucket in use for storing assets. If the bucket doesn't exist,
it creates the bucket, then attempts to set the bucket policy using a
given bucket policy where the s3:GetObject permission is granted for the
given bucket in use.

This task will be run by default by the web application container when
the environment variable MINIO_INIT_BUCKET is set when the web
application container is booted.

Fixes #83.

[1]: https://www.rubydoc.info/gems/paperclip/Paperclip/Storage/S3