Replace poltergeist with selenium-webdriver
Poltergeist (a controller for PhantomJS) has been replaced for system
tests with Selenium-WebDriver. The test helper has been configured to
run system tests that require JS with a headless Chrome.

This means that PhantomJS is no more a dependency for running system
tests. However, this means that Chrome and the ChromeDriver are now
required to run system tests.