Improove testing under Docker
This commit replaces the old docker-compose.yml file with a simple
shellscript that takes care of the stuff that docker-compose.yml can't
deal with, such as stopping and destroying secondary containers when the
RSpec container finishes.

This shell script will spin dependent containers before running the
RSpec container, and will take care of stopping and removing the
dependent containers after the RSpec container finishes.

Additionally, Chrome and ChromeDriver have been installed in the testing
image.