Add a banner when the user sends the DNT header
This commit will display a banner on top of the page whenever the user
browses the page using the DNT header on their browser. To display the
banner, the DNT header must have a value of '1'. It won't be triggered
if the DNT header has a value of '0' or if the DNT header has no value
because the browser does not support it.

If the user closes the banner, the browser will set a cookie to
acknowledge the user has read the banner, in order not to display the
banner anymore on further requests.