Use a compatible way for testing the DNT header
This commit replaces the check for the DNT header. Instead of looking if
the DNT header has a value of "1", it will test if the DNT header has a
value starting with the literal '1'.

This is done because, according to the DNT standard proposal, there may
be DNT extensions in the future after the literal 0 or literal 1.
Therefore, it makes sense to only look for the first character of the
DNT header, respecting future extensions that add extra data to this
header.