chore(lint): lint schema.rb trusting Rails
Ignores db/schema.rb in .rubocop.yml, becase Rubocop and the syntax that
Rails uses to build the db/schema.rb clash on the format. I prefer to
use the Rails way in this file, because it will be touched by Rails
mostly, so it would be a source of conflicts otherwise.