feat: add links (#289)
Links as first class citizens are a complement to video descriptions, allowing videos to point to external URLs. Links have name, description and URL and they appear below the description in the video page.

This PR includes the model, the dashboard controller and views to add, edit and delete links into a video, and the code to present the links in the public webpage.