refactor: remove old data migrator
It's been years since this web application was started, and therefore
there is no point in keeping the data importer to migrate from the
previous WordPress version of this website.

This code is very legacy and it causes issues with some Rubocop offenses
that are difficult to track and fix. Therefore, it is better to just
remove the WordPress importer and maybe work on a backup system
eventually.