fix: paginate the /series page
Fixes https://github.com/makigas/makigas/issues/88

This commit will make the playlists list page not to download every
single card at once, but paginate. The amount of playlists has increased
in the recent years, and thus it is difficult to present all the series
in a single page using a slow connection.