feat: serve JSON variants of the resources
This commit will make the public website to serve requests where the
content negotiation indicates that the user agent would like to receive
an application/json representation of the resources, rather than a
text/html representation.

The intended use case of this feature is JavaScript code that runs in
the text/html variant of the website, to trigger some automated data
fetching in response to user events. Another use case is to allow some
specific user agents to automate access to resources of the website. For
instance, the Clank bot would like to use the JSON version to get
information about videos and to use that information as part of
commands. [1]

[1] https://github.com/makigas/clank/issues/7