feat: order videos and lists by most recent lesson
This commit will make sure that the videos page is sorted by the most
recently published ones. Previous to this commit, they would be still
sorted by the datetime at which they are added to the system, so if you
add older videos after newer ones, they will appear on top.

Also, the list of playlists is now sorted so that playlists most
recently updated (those where a video was added more recently) are
presented first.