fix: replace rails-ujs with @rails/ujs (#595)
It turns out that Rails 6.0 deprecated rails-ujs with @rails/ujs.
Therefore, events for the Rails unobtrusive JS library are not being
handled anymore, which is noticeable because suddenly I'm able to delete
stuff without getting confirmation prompts before.

This commit will uninstall rails-ujs and replace it with @rails/ujs. A
few tests have been added to assert that confirmation prompts are being
presented in destructive actions.