docker: adapt docker-compose to work in Podman
Podman is daemonless and it doesn't run as root, so currently it is
failing to start the containers because they try to write into the host
file system using a root user, which is not allowed.

In the meantime, I change the docker-compose so that it uses local
volumes not exported to the file system, which is good for integration
tests anyway. Do not use this docker-compose.yml in production.