components: migrate to MediaCardComponent
The media card is a component used in the list of topics and in the
video page and it is based in the media component part of Bootstrap 3.
It is used to link to topics and to playlists in the page of a video.

In this commit, this component is migrated to ViewComponent. The
original HTML code is removed from the app/views directory after being
moved into a component that is also tested.