six: refactor metadata into Meta Component
The meta is a little helper component that wraps some basic text with an
icon. At the moment there are only a few properties to customize the
icon and the text to put in the metadata block.

For groups of meta, it is possible to use the meta group component in
order to render the list of metas in a way that formats correctly both
in desktop and in mobile.

This commit also refactors the existing card components so that they use
the new meta group.