tooling: install esbuild
This commit installs esbuild so that running `yarn build` toggles
building all the packs in app/javascript/packs using the esbuild
compiler system instead of webpacker. The packs are compiled in
app/assets/builds for consumption by sprockets.

Note: this commit does not replace at the moment webpacker, but I am not
going to bother with that until I can get esbuild to work as I want.