tooling: use packs generated by esbuild
Instead of using the outcome of webpacker, this commit will replace the
tags used in the HTML templates of the website by those generated when
yarn build is run, so it is now using the packs generated by esbuild.

I have tried to verify that the pages will open, but still haven't
tested using a production build that it will actually compile the packs
where it should. It should be tested via Docker or Podman, yet.