lint: install and configure prettier
Finally, prettier is added to the dist to make sure that the source
codes in the app/javascript directory are always formatted as they
should.

At the moment I am ignoring the app/assets directory, although if I ever
have to touch these files I'll probably lint them as well.