tooling: configure overcommit
Mostly because I'm going to forget about eslint and stylelint, and
I've already forget a couple times about rubocop, it is time to
configure overcommit.

Why overcommit and not husky? I don't know, I have never used husky with
non-JavaScript things, but I know for sure how to use overcommit with
non-Ruby things, so I figured out that it would be faster for me to
setup overcommit.

I do not trust husky, but I do not particularly enjoy overcommit, so I
will setup husky in the future.