six: add a jumbo component
The jumbo is a title enclosed in a dark background to make it look
nicer. It is similar to the jumbotron that used to be present in
Bootstrap, which is what this site previously used to use.

I do not know if it's going to look fine or if I am going to make some
changes to the final layout.