six: address accessibility issues
According to Firefox, there were a couple of inputs without label, a
couple of pictures without an alt image. Most of these images are the
playlist icons, so they can be hidden since they do not convey meaning,
only visual.

Firefox reveals that the old outline system initially implemented does
not actually work, so instead of figuring out how to fix it, I just
replace it with a different kind of outline that actually works and is
actually accessible.