feat: add meta-noindex tag
As I read through the Search Console, it seems that if I really want to
avoid indexing the pages here, I should use meta-noindex rather than
robots.txt.

It turns out that robots.txt is just a guide for navigation, but it is
only adding the noindex meta tag that forces a (good) crawling bot to
actually delete the page from their systems.

What I notice here is that there are still >1000 URLS in the /videos?*
scope that are marked in the Search Console as "not indexed because of a
setting in robots.txt", but it is treated as an error, as if they should
be eventually allowed. I write this change hoping that it stops making
it look like an error on my side and more a "so the owner doesn't want
me to index this".

I still want to keep the robots.txt file to help crawling bots to see
what is important to index, and I want to discourage them from looking
into the /videos? scope.