fix: properly use the lastmod property in sitemap.xml
The sitemap.xml file will now use the last modification date of the
content for videos, playlists and topics, instead of the datetime at
which the sitemap.xml file was generated.

For collection pages like playlists or topics, the last modification
date of the collection will be used (like the last time a playlist was
modified), unless any of the items in the collection is more recent than
the container. In such case, that will be the last modification
timestamp. (Because a title of a video may have changed.)