deps: remove whenever
I am not longer using this to prepare the crontab in production, since I
am using a different strategy for scheduled tasks anyway based on shell
scripts that I configure at the Docker container level. Remove this
dependency because there is a compatibility issue between it and the new
Ruby console anyway.