Show forum threads in the video page
This commit will make the video page pull links to threads from the
forum that are tagged with any of the tags that the video is also tagged
with.

There is a cache to prevent making a lot of requests to the forum,
possibly causing the site to be slow.