Explain to the parrot how to untangle this mess
There are a couple of refactors that are going to receive some help from
an AI agent. Before the work can start, I need to add an AGENT.md file
so that I don't have to explain in every work session how to deal with
the files in my project and provide some global context.