Refactor JSON-LD representations
Rubocop was complaining about the line length of some models so I've
just decided to throw it away and move the JSON-LD representations to
Blueprints.

I have switched the strategy. Rather than returning a simple JSON-LD
object for a page, I am returning a graph. This allows me to decouple
each part of the JSON-LD representation into a separate blueprint, and
then use a primary blueprint to trigger the elements of the graph. It
seems that both Google and Bing support this, and most AI crawlers
should support it too.

Tests are included to check the shapes, both for each node of the
JSON-LD graph, and for the JSON-LD root objects.